[BSE-4469] Split Iceberg IO Code into Multiple Files (#130)
