+++ bodo/tests/test_iceberg/test_iceberg_rest_aws_credentials_provider.cpp
@@ -1,5 +1,5 @@
-#include "../io/_s3_reader.h"
-#include "./test.hpp"
+#include "../../io/_s3_reader.h"
+#include "../test.hpp"
 
 bodo::tests::suite iceberg_rest_aws_credentials_provider_tests([] {
     bodo::tests::test(
