+++ bodo/transforms/distributed_analysis.py
@@ -1456,7 +1456,10 @@ class DistributedAnalysis:
             )
             return
 
-        if fdef == ("iceberg_writer_fetch_theta", "bodo.io.stream_iceberg_write"):
+        if fdef == (
+            "iceberg_writer_fetch_theta",
+            "bodo.io.iceberg.stream_iceberg_write",
+        ):
             # Used to obtain the current value of a theta sketch collection from
             # an Iceberg writer as an array, where each row is the current estimate
             # for that column of the table. Answer is replicated since there is
@@ -1464,7 +1467,7 @@ class DistributedAnalysis:
             _set_REP(self.typemap, self.metadata, self.diag_info, lhs, array_dists)
             return
 
-        if fdef == ("read_puffin_file_ndvs", "bodo.io.stream_iceberg_write"):
+        if fdef == ("read_puffin_file_ndvs", "bodo.io.iceberg.stream_iceberg_write"):
             # Used to the ndvs from a puffin file for testing.
             _set_REP(self.typemap, self.metadata, self.diag_info, lhs, array_dists)
             return
@@ -1476,7 +1479,7 @@ class DistributedAnalysis:
             ),
             (
                 "iceberg_writer_append_table",
-                "bodo.io.stream_iceberg_write",
+                "bodo.io.iceberg.stream_iceberg_write",
             ),
             (
                 "parquet_writer_append_table",
